% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_options.R
\docType{data}
\name{area_options}
\alias{area_options}
\title{Area Options}
\format{
\subsection{\code{area_options}}{

A dataframe with area options, and their long name
\describe{
\item{area_name}{area name}
\item{area_param}{area url parameter}
...
}
}
}
\source{
<gasbuddy.com>
}
\usage{
area_options
}
\description{
Area Options
}
\keyword{datasets}
