% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gas.R
\name{get_gas}
\alias{get_gas}
\title{scapes gasbuddy image pixels and converts to df}
\usage{
get_gas(period = 18, country = "USA", area = "FortCollins")
}
\arguments{
\item{period}{a month number, refer to period_options}

\item{country}{USA or Canada}

\item{area}{this is a country, state, city, check out area_options}
}
\value{
dataframe of prices by date
}
\description{
scapes gasbuddy image pixels and converts to df
}
\examples{
get_gas(18, "USA", "USA")
}
