% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_options.R
\docType{data}
\name{period_options}
\alias{period_options}
\title{Period Options}
\format{
\subsection{\code{period_options}}{

A vector of possible periods
}
}
\source{
<gasbuddy.com>
}
\usage{
period_options
}
\description{
Period Options
}
\keyword{datasets}
